+++ src/ios/OSCamera.swift
@@ -52,8 +52,12 @@ class OSCamera: CDVPlugin {
     @objc(recordVideo:)
     func recordVideo(command: CDVInvokedUrlCommand) {
         self.callbackId = command.callbackId
-        let saveToPhotoAlbum = command.argument(at: 0) as? Bool ?? false
-        let options = OSCAMRVideoOptions(saveToPhotoAlbum: saveToPhotoAlbum)
+        
+        guard let parametersDictionary = command.argument(at: 0) as? [String: Bool],
+              let parametersData = try? JSONSerialization.data(withJSONObject: parametersDictionary),
+              let parameters = try? JSONDecoder().decode(OSCAMRRecordVideoParameters.self, from: parametersData)
+        else { return self.callback(error: .captureVideoIssue) }
+        let options = OSCAMRVideoOptions(from: parameters)
         
         self.commandDelegate.run { [weak self] in
             guard let self = self else { return }
@@ -78,13 +82,37 @@ class OSCamera: CDVPlugin {
         guard let parameterDictionary = command.argument(at: 0) as? [String: Any],
               let parameterData = try? JSONSerialization.data(withJSONObject: parameterDictionary),
               let parameters = try? JSONDecoder().decode(OSCAMRChooseGalleryParameters.self, from: parameterData)
-        else {
-            return self.callback(error: .chooseMultimediaIssue)
-        }
+        else { return self.callback(error: .chooseMultimediaIssue) }
                 
         self.commandDelegate.run { [weak self] in
             guard let self = self else { return }
-            self.plugin?.chooseMultimedia(parameters.mediaType, and: parameters.allowMultipleSelection)
+            self.plugin?.chooseMultimedia(parameters.mediaType, and: parameters.allowMultipleSelection, and: parameters.includeMetadata)
+        }
+    }
+    
+    @objc(playVideo:)
+    func playVideo(command: CDVInvokedUrlCommand) {
+        self.callbackId = command.callbackId
+        
+        guard let parameterDictionary = command.argument(at: 0) as? [String: Any],
+              let parameterData = try? JSONSerialization.data(withJSONObject: parameterDictionary),
+              let parameters = try? JSONDecoder().decode(OSCAMRPlayVideoParameters.self, from: parameterData)
+        else {
+            return self.callback(error: .playVideoIssue)
+        }
+        
+        self.commandDelegate.run { [weak self] in
+            guard let self = self else { return }
+            Task {
+                do {
+                    try await self.plugin?.playVideo(parameters.url)
+                    self.callbackSuccess()
+                } catch let error as OSCAMRError {
+                    self.callback(error: error)
+                } catch {
+                    self.callback(error: .playVideoIssue)
+                }
+            }
         }
     }
 }
@@ -122,5 +150,9 @@ extension OSCamera: OSCAMRCallbackDelegate {
     
     func callback(_ result: String) {
         self.callback(result: result, error: nil)
+    }
+
+    func callbackSuccess() {
+        self.callback("")
     }
 }