+++ src/ios/frameworks/OSCameraLib.xcframework/ios-arm64/OSCameraLib.framework/Headers/OSCameraLib-Swift.h
@@ -195,6 +195,7 @@ typedef unsigned int swift_uint4  __attribute__((__ext_vector_type__(4)));
 #if __has_warning("-Watimport-in-framework-header")
 #pragma clang diagnostic ignored "-Watimport-in-framework-header"
 #endif
+@import AVKit;
 #endif
 
 #pragma clang diagnostic ignored "-Wproperty-attribute-mismatch"
@@ -211,6 +212,11 @@ typedef unsigned int swift_uint4  __attribute__((__ext_vector_type__(4)));
 # pragma clang attribute push(__attribute__((external_source_symbol(language="Swift", defined_in="OSCameraLib",generated_declaration))), apply_to=any(function,enum,objc_interface,objc_category,objc_protocol))
 # pragma pop_macro("any")
 #endif
+
+
+@interface AVPlayerViewController (SWIFT_EXTENSION(OSCameraLib))
+- (void)viewDidDisappear:(BOOL)animated;
+@end